+++ crates/bevy_ui/src/widget/image.rs
@@ -75,7 +75,7 @@ impl Measure for ImageMeasure {
 
 /// Updates content size of the node based on the image provided
 pub fn update_image_content_size_system(
-    mut previous_combined_scale_factor: Local<f64>,
+    mut previous_combined_scale_factor: Local<f32>,
     windows: Query<&Window, With<PrimaryWindow>>,
     ui_scale: Res<UiScale>,
     textures: Res<Assets<Image>>,
@@ -105,7 +105,7 @@ pub fn update_image_content_size_system(
                 image_size.size = size;
                 content_size.set(ImageMeasure {
                     // multiply the image size by the scale factor to get the physical size
-                    size: size * combined_scale_factor as f32,
+                    size: size * combined_scale_factor,
                 });
             }
         }
@@ -116,7 +116,7 @@ pub fn update_image_content_size_system(
 
 /// Updates content size of the node based on the texture atlas sprite
 pub fn update_atlas_content_size_system(
-    mut previous_combined_scale_factor: Local<f64>,
+    mut previous_combined_scale_factor: Local<f32>,
     windows: Query<&Window, With<PrimaryWindow>>,
     ui_scale: Res<UiScale>,
     atlases: Res<Assets<TextureAtlas>>,
@@ -153,7 +153,7 @@ pub fn update_atlas_content_size_system(
                 image_size.size = size;
                 content_size.set(ImageMeasure {
                     // multiply the image size by the scale factor to get the physical size
-                    size: size * combined_scale_factor as f32,
+                    size: size * combined_scale_factor,
                 });
             }
         }
